// api/generated/model/exerciseResponseDto.ts
/**
 * Generated by orval v8.2.0 🍺
 * Do not edit manually.
 * LgymApi.Api
 * OpenAPI spec version: 1.0
 */
import type { EnumLookupDto } from './enumLookupDto';

export interface ExerciseResponseDto {
  /** @nullable */
  _id?: string | null;
  /** @nullable */
  name?: string | null;
  /** @nullable */
  user?: string | null;
  bodyPart?: EnumLookupDto;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  image?: string | null;
}